Fix responsiveness
