+++ client/package.json
@@ -108,6 +108,7 @@
 		"@wordpress/is-shallow-equal": "^5.22.0",
 		"@wordpress/private-apis": "^1.22.0",
 		"@wordpress/react-i18n": "^4.22.0",
+		"@wordpress/viewport": "^6.22.0",
 		"@wordpress/warning": "^3.22.0",
 		"autosize": "^4.0.4",
 		"body-parser": "^1.20.3",
