+++ client/signup/steps/domains/index.jsx
@@ -14,6 +14,7 @@ import {
 } from '@automattic/onboarding';
 import { withShoppingCart } from '@automattic/shopping-cart';
 import { getQueryArg } from '@wordpress/url';
+import { withViewportMatch } from '@wordpress/viewport';
 import clsx from 'clsx';
 import { localize } from 'i18n-calypso';
 import { defer, get, isEmpty } from 'lodash';
@@ -92,10 +93,14 @@ import { setDesignType } from 'calypso/state/signup/steps/design-type/actions';
 import { getDesignType } from 'calypso/state/signup/steps/design-type/selectors';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
 import DomainsMiniCart from './domains-mini-cart';
-import { getExternalBackUrl, shouldUseMultipleDomainsInCart } from './utils';
+import {
+	getExternalBackUrl,
+	shouldUseMultipleDomainsInCart,
+	sortProductsByPriceDescending,
+} from './utils';
 import './style.scss';
 
-export class RenderDomainsStep extends Component {
+class RenderDomainsStepComponent extends Component {
 	static propTypes = {
 		cart: PropTypes.object,
 		shoppingCartManager: PropTypes.any,
@@ -745,7 +750,7 @@ export class RenderDomainsStep extends Component {
 						return ! this.state.domainsWithMappingError.includes( product.meta );
 					} );
 					// Sort products to ensure the user gets the best deal with the free domain bundle promotion.
-					const sortedProducts = this.sortProductsByPriceDescending( productsInCart );
+					const sortedProducts = sortProductsByPriceDescending( productsInCart );
 					this.props.shoppingCartManager
 						.replaceProductsInCart( sortedProducts )
 						.then( () => {
@@ -784,29 +789,6 @@ export class RenderDomainsStep extends Component {
 		}
 
 		this.setState( { isCartPendingUpdateDomain: null } );
-	}
-
-	sortProductsByPriceDescending( productsInCart ) {
-		// Sort products by price descending, considering promotions.
-		const getSortingValue = ( product ) => {
-			if ( product.item_subtotal_integer !== 0 ) {
-				return product.item_subtotal_integer;
-			}
-
-			// Use the lowest non-zero new_price or fallback to item_original_cost_integer.
-			const nonZeroPrices =
-				product.cost_overrides
-					?.map( ( override ) => override.new_price * 100 )
-					.filter( ( price ) => price > 0 ) || [];
-
-			return nonZeroPrices.length
-				? Math.min( ...nonZeroPrices )
-				: product.item_original_cost_integer;
-		};
-
-		return productsInCart.sort( ( a, b ) => {
-			return getSortingValue( b ) - getSortingValue( a );
-		} );
 	}
 
 	removeDomainClickHandler = ( domain ) => async () => {
@@ -1010,6 +992,7 @@ export class RenderDomainsStep extends Component {
 		const content = [
 			domainsInCart.length > 0 || this.state.wpcomSubdomainSelected ? (
 				<DomainsMiniCart
+					isMobile={ ! this.props.isDesktop }
 					domainsInCart={ domainsInCart }
 					domainRemovalQueue={ this.state.domainRemovalQueue }
 					cartIsLoading={ cartIsLoading }
@@ -1576,6 +1559,10 @@ export class RenderDomainsStep extends Component {
 		);
 	}
 }
+
+export const RenderDomainsStep = withViewportMatch( { isDesktop: '>= large' } )(
+	RenderDomainsStepComponent
+);
 
 export const submitDomainStepSelection = ( suggestion, section ) => {
 	let domainType = 'domain_reg';
