+++ client/signup/steps/domains/utils.js
@@ -38,3 +38,24 @@ export function shouldUseMultipleDomainsInCart( flowName ) {
 
 	return enabledFlows.includes( flowName );
 }
+
+export function sortProductsByPriceDescending( productsInCart ) {
+	// Sort products by price descending, considering promotions.
+	const getSortingValue = ( product ) => {
+		if ( product.item_subtotal_integer !== 0 ) {
+			return product.item_subtotal_integer;
+		}
+
+		// Use the lowest non-zero new_price or fallback to item_original_cost_integer.
+		const nonZeroPrices =
+			product.cost_overrides
+				?.map( ( override ) => override.new_price * 100 )
+				.filter( ( price ) => price > 0 ) || [];
+
+		return nonZeroPrices.length ? Math.min( ...nonZeroPrices ) : product.item_original_cost_integer;
+	};
+
+	return productsInCart.sort( ( a, b ) => {
+		return getSortingValue( b ) - getSortingValue( a );
+	} );
+}
